# api/api-docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AgDx API",
    "description": "AgDx API facilitates sharing of crop pest and disease data between platforms and/or by providing links between complementary ICT tools where appropriate.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0.0-beta1"
  },
...
  "servers": [
    {
      "url": "http://localhost:9100",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v1/pest/update/{id}": {
      "put": {
        "tags": [
          "Pest data"
        ],
        "summary": "Update pest record",
        "operationId": "updatePestData",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PestEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PestEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/crop/update/{id}": {
      "put": {
        "tags": [
          "Crop data"
        ],
        "summary": "Update crop record",
        "operationId": "updateCropData",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/crop/pest/update/{id}": {
      "put": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Update crop pest record",
        "operationId": "updatePestData_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropPestEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropPestEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/crop/disease/update/{id}": {
      "put": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Update crop disease record",
        "operationId": "updatePestData_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropDiseaseEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropDiseaseEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/pest/add": {
      "post": {
        "tags": [
          "Pest data"
        ],
        "summary": "Add new pest",
        "operationId": "addPestData",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PestEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PestEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/crop/pest/add": {
      "post": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Add new pest",
        "operationId": "addPestData_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropPestEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropPestEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/crop/disease/add": {
      "post": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Add new crop pest record",
        "operationId": "addPestData_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropDiseaseEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropDiseaseEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/crop/add": {
      "post": {
        "tags": [
          "Crop data"
        ],
        "summary": "Add new crop",
        "operationId": "addCropData",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
...
    "/api/v1/crop/{cropName}/scientific": {
      "get": {
        "tags": [
          "Crop data"
        ],
        "summary": "Get crops using scientific name",
        "operationId": "getCropByScientificName",
        "parameters": [
          {
            "name": "cropName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/{cropName}/common": {
      "get": {
        "tags": [
          "Crop data"
        ],
        "summary": "Get crops using common name",
        "operationId": "getCropByCommonName",
        "parameters": [
          {
            "name": "cropName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/crop/pest/admin-level/{adminName}/two": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Find crop pest by administration levels",
        "operationId": "getDiseaseByAdminLevelTwo",
        "parameters": [
          {
            "name": "adminName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/admin-level/{adminName}/three": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Find crop pest by administration levels",
        "operationId": "getDiseaseByAdminLevelThree",
        "parameters": [
          {
            "name": "adminName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/admin-level/{adminName}/one": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Find crop pest by administration levels",
        "operationId": "getDiseaseByAdminLevelOne",
        "parameters": [
          {
            "name": "adminName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/crop/disease/{diseaseName}/scientific": {
      "get": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Get details of specific disease using its scientific name",
        "operationId": "getDiseaseByScientificName",
        "parameters": [
          {
            "name": "diseaseName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropDiseaseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/disease/{diseaseName}/common": {
      "get": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Get details of specific disease using its common name",
        "operationId": "getDiseaseByCommonName",
        "parameters": [
          {
            "name": "diseaseName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropDiseaseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/disease/verification-method/{verificationMethod}": {
      "get": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Find crop disease by verification methods",
        "operationId": "getDiseaseByVerificationMethod",
        "parameters": [
          {
            "name": "verificationMethod",
            "in": "path",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropDiseaseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/disease/reporting-date": {
      "get": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Find crop disease by reporting date range",
        "operationId": "getDiseaseByReportingDate",
        "parameters": [
          {
            "name": "fromDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropDiseaseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/disease/record-date": {
      "get": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Find crop disease by recording date range",
        "operationId": "getDiseaseByRecordingDate",
        "parameters": [
          {
            "name": "fromDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropDiseaseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/disease/project/{project}": {
      "get": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Get list of diseases by project name",
        "operationId": "getDiseaseByProjectName",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropDiseaseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/disease/pathogen/{name}/scientific": {
      "get": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Get list of diseases by scientific pathogen name ",
        "description": "Search parameters are optional",
        "operationId": "getDiseaseByScientificPathogen",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropDiseaseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/disease/pathogen/{name}/common": {
      "get": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Get list of diseases by common pathogen names",
        "description": "Search parameters are optional",
        "operationId": "getDiseaseByCommonPathogen",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropDiseaseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/disease/observe-date": {
      "get": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Find crop disease by observation date range",
        "operationId": "getDiseaseByObservedDate",
        "parameters": [
          {
            "name": "fromDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropDiseaseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/disease/genebank/{accessionNumber}": {
      "get": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Get list of diseases by genebank number",
        "description": "Search can be partial",
        "operationId": "getDiseaseByGeneBankNumber",
        "parameters": [
          {
            "name": "accessionNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropDiseaseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/disease/cultivar/{cultivar}": {
      "get": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Find crop disease by cultivar type",
        "operationId": "getDiseaseByCultivarName",
        "parameters": [
          {
            "name": "cultivar",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropDiseaseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/disease/crop-name/{cropName}": {
      "get": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Find crop disease by crop name",
        "operationId": "getDiseaseByCropName",
        "parameters": [
          {
            "name": "cropName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropDiseaseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/disease/country/{countryCode}": {
      "get": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Find crop diseases by country code",
        "operationId": "getDiseaseByCountry",
        "parameters": [
          {
            "name": "countryCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropDiseaseEntity"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PestEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "commonName": {
            "type": "string",
            "description": "Pest common name",
            "example": "Wooly apple aphid"
          },
          "scientificName": {
            "type": "string",
            "description": "Pest scientific name",
            "example": "Eriosoma lanigerum"
          },
          "order": {
            "type": "string",
            "description": "Pest order",
            "example": "Hemiptera"
          },
          "family": {
            "type": "string",
            "description": "Pest family",
            "example": "Aphididae"
          },
          "genus": {
            "type": "string",
            "description": "Pest genus",
            "example": "Eriosoma"
          },
          "category": {
            "type": "string",
            "description": "Pest family",
            "example": "Insect",
            "enum": [
              "Virus",
              "Bacteria",
              "Nematode",
              "Plant",
              "Insect"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "api": {
        "type": "http",
        "in": "header",
        "scheme": "basic"
      }
    }
  }
}